Check for existing GIT repo
This is to avoid accidential data loss during `composer update`

+++ _bluespice/pre-autoload-dump.d/05-add_installer_overrides.sh
@@ -3,6 +3,11 @@
 # Copyright: 2021
 # License: GPLv3
 
+GREEN='\033[0;32m'
+RED='\033[0;31m'
+PURPLE='\033[0;35m'
+NC='\033[0m'
+
 BRANCH=$(git branch | cut -f 2 -d " ")
 BRANCH=${BRANCH%%[[:space:]]}
 BRANCH=${BRANCH##[[:space:]]}
@@ -13,6 +18,14 @@ then
 	BRANCH="REL1_35"
 fi
 
-rm -rf mw-config/overrides
-git clone -b $BRANCH --depth 1 https://gerrit.wikimedia.org/r/bluespice/mw-config/overrides mw-config/overrides
+printf "\n${PURPLE}Fetching installer: ${NC}"
 
+if ! [ -d "mw-config/overrides/.git" ]
+then
+	rm -rf mw-config/overrides
+	git clone -b $BRANCH --depth 1 https://gerrit.wikimedia.org/r/bluespice/mw-config/overrides mw-config/overrides
+else
+	git -C mw-config/overrides/ pull
+fi
+
+printf "${GREEN}[ DONE ]${NC}\n"
